+++ packages/llm-bridge-spec/src/index.ts
@@ -1,1 +1,3 @@
-export * from './types';
+export * from './bridge/types';
+export * from './message/types';
+export * from './manifest/types';
